+++ tests/t00028/test_case.h
@@ -54,6 +54,12 @@ note.)";
     REQUIRE_THAT(puml, HasNote(_A("F"), "bottom", "F enum note."));
     REQUIRE_THAT(puml, !HasNote(_A("G"), "left", "G class note."));
     REQUIRE_THAT(puml, HasNote(_A("R"), "right", "R class note."));
+    REQUIRE_THAT(puml,
+        HasMemberNote(_A("R"), "aaa", "left", "R contains an instance of A."));
+    REQUIRE_THAT(
+        puml, !HasMemberNote(_A("R"), "bbb", "right", "R class note."));
+    REQUIRE_THAT(
+        puml, HasMemberNote(_A("R"), "ccc", "left", "Reference to C."));
 
     save_puml(config.output_directory() + "/" + diagram->name + ".puml", puml);
 }